requirements.txt
# Face Concern Detector - Requirements
# Optimized for Mac M1/M2 with dual dataset support

# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Computer Vision & Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Face Detection & Recognition  
mtcnn>=0.1.1
facenet-pytorch>=2.5.3

# Explainable AI
grad-cam>=1.4.8

# Dataset Download & Processing
kagglehub>=0.2.0
requests>=2.31.0

# Scientific Computing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Flask API (Production Deployment)
flask>=2.3.0
flask-cors>=4.0.0
werkzeug>=2.3.0

# Development & Jupyter
jupyter>=1.0.0
ipywidgets>=8.0.0
notebook>=7.0.0

# Optional: Enhanced functionality
# albumentations>=1.3.0  # Advanced augmentations
# tensorboard>=2.14.0    # Training visualization
# wandb>=0.15.0          # Experiment tracking
